refactor(mux): colocate topology cache policy and protocol
